+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Added
   * AZ-174: Add Availability Zone option 
   * AZ-173: Allow to configure disk size and type
+  * AZ-108: Linux-VM enable diagnostics
 
 # v2.1.0 - 2020-01-07